+++ src/index.ts
@@ -1,0 +1,7 @@
+export * from '@sanity/block-content-to-hyperscript'
+export * from 'types'
+export {defaultSerializers} from 'targets/dom'
+import BlockContent from 'BlockContent'
+export * from 'BlockContent'
+
+export default BlockContent
